✨ Mod Support for Exposure (1.21+) - Experimental
Experimental Mod Support, some of these assets of the Exposure mod is not final and isn't faithful to it's respective Default Dark Mode style compared to other mods I've supported. Exposure is one of the few mods that has GUIs that are harder to work with when it comes to editing it for Default Dark Mode: Expansion.