Update pack.mcmeta (version fix)
In the GitHub version it's defined as "2025.X.X" because this version of the resource pack also contains assets licensed by NebuIr and isn't considered an actual version of DDM, use this ONLY for personalization and if you wanted both DDM and DDME assets combined. versions under "2025.X.X" can not be redistributed.